identity/jwt_validator.py
"""
===============================================================================
MODULE: jwt_validator.py
===============================================================================

PURPOSE:
    Validates JWT tokens from Cognito.

WHEN TO USE THIS MODULE:
    - Token validation: Verify JWT tokens
    - Authentication: Check token validity

USAGE EXAMPLES:
    from identity.jwt_validator import validate_token
    
    claims = validate_token(token)
    if claims:
        user_id = claims['sub']

WHAT THIS MODULE DOES:
    1. Validates JWT token signature
    2. Checks token expiration
    3. Verifies token issuer
    4. Extracts user claims

RELATED FILES:
    - auth/cognito_client.py - Uses token validation
    - identity/cognito_integration.py - Cognito integration

AUTHOR: Enterprise Cloud Engineer Agent Project
DATE: 2025-01-XX
VERSION: 1.0.0
===============================================================================
"""

import os
import jwt
import requests
from typing import Optional, Dict, Any
from functools import lru_cache

from utils.logging_config import setup_logger

logger = setup_logger(__name__)


@lru_cache(maxsize=1)
def get_jwks() -> Dict[str, Any]:
    """
    Get JWKS (JSON Web Key Set) from Cognito.
    
    RETURNS:
        Dict[str, Any]: JWKS
    """
    pool_id = os.getenv('COGNITO_USER_POOL_ID')
    region = os.getenv('AWS_REGION', 'us-east-2')
    
    if not pool_id:
        return {}
    
    jwks_url = f"https://cognito-idp.{region}.amazonaws.com/{pool_id}/.well-known/jwks.json"
    
    try:
        response = requests.get(jwks_url, timeout=5)
        response.raise_for_status()
        return response.json()
    except Exception as e:
        logger.error(f"Failed to fetch JWKS: {e}")
        return {}


def validate_token(token: str) -> Optional[Dict[str, Any]]:
    """
    Validate JWT token.
    
    ARGUMENTS:
        token (str): JWT token to validate
    
    RETURNS:
        Optional[Dict[str, Any]]: Token claims if valid, None otherwise
    """
    try:
        # Decode token without verification first to get header
        unverified = jwt.decode(token, options={"verify_signature": False})
        
        # Get JWKS
        jwks = get_jwks()
        if not jwks:
            logger.warning("JWKS not available, skipping signature verification")
            return unverified
        
        # TODO: Implement full token validation with JWKS
        # For now, return decoded token
        return unverified
    
    except jwt.ExpiredSignatureError:
        logger.error("Token expired")
        return None
    except jwt.InvalidTokenError as e:
        logger.error(f"Invalid token: {e}")
        return None
    except Exception as e:
        logger.error(f"Token validation error: {e}")
        return None
